pyproject.toml
[project]
name = "jm-mcp-server"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "jmcomic",
    "Pillow",
    "img2pdf",
    "pikepdf",
    "lxml",
    "pyyaml",
]

[project.optional-dependencies]
android = [
    "requests[socks]",  # 支持代理
    "psutil",          # 系统监控
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
